Create MultiDimensionalSearch
Abstract:

Implement Multi-Dimensional search. Each user carry tens of thousands of products, and each product has many attributes (Name, Size, Description, Keywords, Manufacturer, Price, etc.).The web site allows users to specify attributes of products that they are seeking, and displays products that have most of those attributes. To make search efficient, the data is organized using appropriate data structures, such as balanced trees and hashing.  But, if products are organized by Name, search by price must work efficiently. The solution, called indexing in databases, is to create a new set of references to the objects for each search field, and organize them to implement search operations on that field efficiently. As the objects change, these access structures have to be kept consistent.

Problem Statement:

In this project, each customer object has 3 attributes: id (long int), categories (one or more integers 1-999), and amount (dollars and cents).The amount field stores the total revenue that the customer has generated for our company.  The categories field is a set of department codes of our company (encoded as an integer in [1,999]) that stock items of interest to the customer.  Customers are uniquely identified by their id (key field). Need to write a java Program 


Primary Functions implemented and methodologies followed:

insert(id,categories[]) : 

Data structures used: Tree Map with key as id of customer and value as Customer node. A hash map data structure is also maintained to do a search for Top three. 

All the customers are stored in a tree map with key value as the id of the customer. 

find(id): return amount spent by customer until now (or -1, if  such a customer does not exist).

Data structures used: The tree map is searched for a particular customer. 

delete(id): delete customer's records from storage.  Returns the amount field of the deleted customer (or -1, if such an id did not exist).

Data structures used: The tree map which contains the customer ids. Customer is searched based on id and deleted.  

TopThree(category): given a category k, find the top three customers (in terms of amount spent) who are interested in category k.

Data structures used: Hash Map which has key as category and value as a tree set of nodes of all customers who have that category.

Nodes in tree set is arranged in the decreasing order of amounts spent by customers. First three nodes in the tree set is returned as top three.

AddInterests(id,categories): Add new interests to the list of  a customer's categories.  Some of them may already be in the list of categories of this customer.  Return the number of new  categories added to that customer's record.

Data structures used: Tree map with key as id of customer and value as Customer node , Hash map which has key as category and value as a tree set of nodes of all customers who have that category.

RemoveInterests(id, categories): Remove some categories from the list of categories associated with a customer.  Return the  number of categories left in the customer's record.
 
Data structures used: Tree map with key as id of customer and value as Customer node , Hash map which has key as category and value as a tree set of nodes of all customers who have that category.


AddRevenue(id, purchase): update customer record by adding a  purchase amount spent by a customer on our company's product.
      Returns the net amount spent by the customer after adding this  purchase, truncated to just dollars

Data structures used: Tree map with key as id of customer and value as Customer node , Hash map which has key as category and value as a tree set of nodes of all customers who have that category.

Each time when this method is called, the node from the tree set is removed, node is updated and later added back to tree set inside the hash map.

Range(low, high): number of customers whose amount is at least "low" and at most "high".

Data structures used: A tree set which has nodes of customers arranged based on decreasing value of their amounts.  

Tree set is iterated from high to low to output the number of customers between that range and their count is taken.

SameSame(): Find customers who have exactly the same set of 5 or   more categories of interest.  Creative solutions that are elegant and efficient will be awarded excellence credit. Return the number of distinct customers who have exactly the  same set of 5 or more interests as another customer.

Data structures used: Hash map with key as string of combination of categories and value as count of customers having this combination. Customers having 5 or more than 5 categories are only considered to be stored inside this has map.

Hash map is iterated linearly to fetch all values with count more than one which is the return value.

NumberPurchases(id): The number of times customer has purchased   products (i.e., number of calls to AddRevenue for this customer). Return -1 if no such customer exists.

Data structures used: Node class has a variable which maintains the count of the number of calls to Add revenue.